+++ view/Hex.java
@@ -1,4 +1,4 @@
-package model;
+package view;
 
 import java.awt.BorderLayout;
 
@@ -8,6 +8,8 @@ import javax.swing.JMenu;
 import javax.swing.JMenuBar;
 import javax.swing.JMenuItem;
 import javax.swing.SwingUtilities;
+
+import model.HexModel;
 
 public class Hex {
